Improved forms validation, fixed styling and naming
