+++ src/app/login/login.component.html
@@ -2,32 +2,30 @@
   <h2 class="mat-display-1">Login Form</h2>
 
   <mat-form-field>
-    <input matInput placeholder="Email" formControlName="email" minlength="4" maxlength="20">
-
-    <mat-error *ngIf="form.email.invalid && form.email.touched">
-      Email is missing or invalid
+    <input matInput placeholder="Email" formControlName="email">
+    <mat-hint align="end">
+      Format: johndoe@mail.com
+    </mat-hint>
+    <mat-error *ngIf="isFieldInvalid('email')">
+      Please enter valid Email (format: johndoe@mail.com)
     </mat-error>
-
   </mat-form-field>
 
   <mat-form-field>
-    <input matInput placeholder="Password" type="password" formControlName="password" minlength="6" maxlength="16">
-
+    <input matInput placeholder="Password" type="password" formControlName="password">
     <mat-hint align="end">
       Between 6 and 16 symbols, must have at least one letter and one number
     </mat-hint>
-    <mat-error *ngIf="form.password.invalid && form.password.touched">
-      Password is missing or invalid
+    <mat-error *ngIf="isFieldInvalid('password')">
+      Please enter valid Password (6-16 symbols, 1 digit and 1 letter at least)
     </mat-error>
-
   </mat-form-field>
-
 
   <div class="div-buttons">
     <button type="submit" mat-raised-button [disabled]="loginForm.invalid">
       Log In
     </button>
-    <button mat-raised-button (click)="goToRegistrationPage()">
+    <button type="button" mat-raised-button (click)="goToRegistrationPage()">
       Create New Account
     </button>
   </div>
