+++ src/app/login/login.component.scss
@@ -1,7 +1,7 @@
 .container {
   display: flex;
   flex-direction: column;
-  max-width: 600px;
+  max-width: 37rem;
   margin: 5rem auto;
 
   h2 {
