+++ src/app/login/login.interface.ts
@@ -1,4 +1,4 @@
-export interface User {
+export interface Login {
   email: string;
   password: string;
 }
