+++ src/app/login/login.module.ts
@@ -7,8 +7,6 @@ import { MatButtonModule } from '@angular/material/button';
 import { LoginComponent } from './login.component';
 import { HttpClientModule } from '@angular/common/http';
 
-
-
 @NgModule({
   declarations: [LoginComponent],
   imports: [
