+++ src/app/registration-user/registration-user.interface.ts
@@ -1,9 +1,8 @@
-import { User } from './../login/login.interface';
-
-export interface NewUser extends User {
+export interface NewUser {
   firstName: string;
   lastName: string;
   phone: string;
+  email: string;
   password: string;
   id?: number
 }
