--- src/app/shared/services/user/user.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { UserService } from './user.service';
-
-xdescribe('UserService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  xit('should be created', () => {
-    const service: UserService = TestBed.get(UserService);
-    expect(service).toBeTruthy();
-  });
-});
